python-js/requirements.txt
beautifulsoup4==4.14.2
cryptography==46.0.2
Django==5.2.7
django_oidc_provider==0.8.4
django_sslserver==0.22
face_recognition==1.3.0
fido2==2.0.0
Flask==3.1.2
Flask_Bcrypt==1.0.1
Flask-Session==0.8.0
flask_cors==6.0.1
numpy==2.3.3
Pillow==11.3.0
PyJWT==2.10.1
pyotp==2.9.0
pysaml2==7.5.2
qrcode==8.2
Requests==2.32.5
